requirements.txt
py-cord
python-dotenv
discord-py-interactions
